Introduced Dataset class
